Move RandomInputStream and its test to java-manta-it and make -benchmark depend on the unshaded module
