+++ java-manta-it/pom.xml
@@ -38,13 +38,6 @@
         </dependency>
 
         <dependency>
-            <groupId>com.joyent.manta</groupId>
-            <artifactId>java-manta-benchmark</artifactId>
-            <version>3.1.7-SNAPSHOT</version>
-            <scope>test</scope>
-        </dependency>
-
-        <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-collections4</artifactId>
             <version>${dependency.commons-collections.version}</version>
