+++ java-manta-it/src/test/java/com/joyent/test/util/RandomInputStreamTest.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.joyent.manta.benchmark;
+package com.joyent.test.util;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
